requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
defusedxml==0.7.1
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
lizard==1.17.10
oauthlib==3.2.2
packaging==23.0
PyDriller==2.4
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
smmap==5.0.0
tqdm==4.64.1
types-pytz==2022.7.1.0
typing_extensions==4.4.0
urllib3==1.26.14
